backend/pyproject.toml
[project]
name = "code-compass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "future>=1.0.0",
    "llama-index>=0.12.44",
    "llama-index-core>=0.12.44",
    "llama-index-embeddings-azure-openai>=0.3.8",
    "llama-index-llms-azure-openai>=0.3.4",
    "numpy>=1.26.4",
    "scipy>=1.12.0",
    "uvicorn[standard]>=0.34.3",
    "tree-sitter>=0.24.0",
    "tree-sitter-language-pack>=0.8.0",
    "ipykernel>=6.29.5",
    "llama-index-vector-stores-milvus>=0.8.5",
]
